Update account view to always show Add button
#38
