+++ src/layout/Account.tsx
@@ -1,9 +1,6 @@
-import React, {useContext} from 'react';
-import {ChainApiContext} from 'context/ChainApiContext';
-import {getAccountsIdentityInfo} from 'service/api/account';
 import {AccountId, Registration} from '@polkadot/types/interfaces';
 import {IdentityInfo} from '@polkadot/types/interfaces/identity/types';
-import {useQuery} from 'react-query';
+import {useAccountIdentity} from 'src/hook/useAccountIdentity';
 
 export function Account({
   id,
@@ -23,21 +20,4 @@ export function Account({
   }
 
   return children(account);
-}
-
-export function useAccountIdentity(id?: string) {
-  const {api} = useContext(ChainApiContext);
-
-  return useQuery(
-    ['account_identity', id],
-    async () => {
-      if (!id) {
-        return undefined;
-      }
-
-      const accounts = api ? await getAccountsIdentityInfo([id], api) : undefined;
-      return accounts?.[0];
-    },
-    {staleTime: 1000 * 60, enabled: !!id},
-  );
 }
